+++ src/graph/data/repeat.rs
@@ -5,7 +5,7 @@ impl Tensor {
     // apply function across every node --> concat
     // technically, there is a permute at the end, so
     // every node --> apply func --> permute to last dim --> concat last dim --> permute to orig dim
-    pub fn repeat<F> (&self, f: F, len: usize, dim: usize) -> Tensor 
+    pub fn repeat<F> (&self, f: F, len: usize, dim: i32) -> Tensor 
         where F: Fn(&Tensor, usize) -> Tensor
     {
         let mut vector_nodes: Vec<Tensor> = vec![];
